# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FastapMenu - Digital Restaurant Dining Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{restaurantId}/tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List tables
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create table
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /restaurants/{restaurantId}/tables/{tableId}:
    put:
      operationId: updateTable
      tags: [tables]
      summary: Update table
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete table
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── MENU CATEGORIES ─────────────────────────────────────────────────────────
...
  /restaurants/{restaurantId}/orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /restaurants/{restaurantId}/orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── STAFF ───────────────────────────────────────────────────────────────────
...
  /public/orders:
    post:
      operationId: placeOrder
      tags: [public]
      summary: Place an order (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /public/reservations:
    post:
      operationId: createReservation
      tags: [public]
      summary: Create reservation (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /public/feedback:
    post:
      operationId: submitFeedback
      tags: [public]
      summary: Submit feedback (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    Table:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        branchId:
          type: ["integer", "null"]
        name:
          type: string
        zone:
          type: ["string", "null"]
        capacity:
          type: integer
        qrCodeUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, restaurantId, name, capacity, isActive, createdAt]
    TableInput:
      type: object
      properties:
        name:
          type: string
        branchId:
          type: ["integer", "null"]
        zone:
          type: ["string", "null"]
        capacity:
          type: integer
        isActive:
          type: boolean
      required: [name, capacity]

    # ─── CATEGORY ─────────────────────────────────────────────────────────────
...
    Order:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        tableId:
          type: ["integer", "null"]
        tableName:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        type:
          type: string
          enum: [dine_in, delivery, takeaway]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, completed, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        createdAt:
          type: string
          format: date-time
      required: [id, restaurantId, type, status, items, subtotal, tax, total, paymentStatus, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        variant:
          type: ["string", "null"]
        addons:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        subtotal:
          type: number
      required: [id, menuItemId, name, price, quantity, addons, subtotal]
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, completed, cancelled]
      required: [status]
    PlaceOrderInput:
      type: object
      properties:
        restaurantId:
          type: integer
        tableId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        type:
          type: string
          enum: [dine_in, delivery, takeaway]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlaceOrderItemInput"
        notes:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
      required: [restaurantId, type, items]
    PlaceOrderItemInput:
      type: object
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        variant:
          type: ["string", "null"]
        addons:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
      required: [menuItemId, quantity]
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: string
        guestCount:
          type: integer
        tableId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, restaurantId, customerName, customerPhone, date, time, guestCount, status, createdAt]
    ReservationInput:
      type: object
      properties:
        restaurantId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: string
        guestCount:
          type: integer
        notes:
          type: ["string", "null"]
      required: [restaurantId, customerName, customerPhone, date, time, guestCount]
...
    Feedback:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        orderId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        rating:
          type: integer
        foodRating:
          type: ["integer", "null"]
        serviceRating:
          type: ["integer", "null"]
        ambienceRating:
          type: ["integer", "null"]
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, restaurantId, rating, createdAt]
    FeedbackInput:
      type: object
      properties:
        restaurantId:
          type: integer
        orderId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        rating:
          type: integer
        foodRating:
          type: ["integer", "null"]
        serviceRating:
          type: ["integer", "null"]
        ambienceRating:
          type: ["integer", "null"]
        comment:
          type: ["string", "null"]
      required: [restaurantId, rating]

    # ─── WAITER CALL ──────────────────────────────────────────────────────────
...